third_party/WebKit/LayoutTests/external/wpt/html/dom/reflection-metadata-expected.txt

This is a testharness.js-based test.
PASS head.title: 32 tests
PASS head.lang: 32 tests
PASS head.dir: 62 tests
PASS head.className (<head class>): 32 tests
PASS head.hidden: 33 tests
PASS head.accessKey: 32 tests
PASS head.tabIndex: 24 tests
PASS title.title: 32 tests
PASS title.lang: 32 tests
PASS title.dir: 62 tests
PASS title.className (<title class>): 32 tests
PASS title.hidden: 33 tests
PASS title.accessKey: 32 tests
PASS title.tabIndex: 24 tests
PASS base.title: 32 tests
PASS base.lang: 32 tests
PASS base.dir: 62 tests
PASS base.className (<base class>): 32 tests
PASS base.hidden: 33 tests
PASS base.accessKey: 32 tests
PASS base.tabIndex: 24 tests
PASS base.target: 32 tests
PASS link.title: 32 tests
PASS link.lang: 32 tests
PASS link.dir: 62 tests
PASS link.className (<link class>): 32 tests
PASS link.hidden: 33 tests
PASS link.accessKey: 32 tests
PASS link.tabIndex: 24 tests
PASS link.href: 38 tests
PASS link.crossOrigin: 52 tests
PASS link.rel: 32 tests
PASS link.media: 32 tests
FAIL link.nonce: typeof IDL attribute assert_equals: expected "string" but got "undefined"
FAIL link.nonce: IDL get with DOM attribute unset assert_equals: expected (string) "" but got (undefined) undefined
FAIL link.nonce: setAttribute() to "" assert_equals: IDL get expected (string) "" but got (undefined) undefined
FAIL link.nonce: setAttribute() to " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo " assert_equals: IDL get expected (string) " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo " but got (undefined) undefined
FAIL link.nonce: setAttribute() to undefined assert_equals: IDL get expected (string) "undefined" but got (undefined) undefined
FAIL link.nonce: setAttribute() to 7 assert_equals: IDL get expected (string) "7" but got (undefined) undefined
FAIL link.nonce: setAttribute() to 1.5 assert_equals: IDL get expected (string) "1.5" but got (undefined) undefined
FAIL link.nonce: setAttribute() to true assert_equals: IDL get expected (string) "true" but got (undefined) undefined
FAIL link.nonce: setAttribute() to false assert_equals: IDL get expected (string) "false" but got (undefined) undefined
FAIL link.nonce: setAttribute() to object "[object Object]" assert_equals: IDL get expected (string) "[object Object]" but got (undefined) undefined
FAIL link.nonce: setAttribute() to NaN assert_equals: IDL get expected (string) "NaN" but got (undefined) undefined
FAIL link.nonce: setAttribute() to Infinity assert_equals: IDL get expected (string) "Infinity" but got (undefined) undefined
FAIL link.nonce: setAttribute() to -Infinity assert_equals: IDL get expected (string) "-Infinity" but got (undefined) undefined
FAIL link.nonce: setAttribute() to "\0" assert_equals: IDL get expected (string) "\0" but got (undefined) undefined
FAIL link.nonce: setAttribute() to null assert_equals: IDL get expected (string) "null" but got (undefined) undefined
FAIL link.nonce: setAttribute() to object "test-toString" assert_equals: IDL get expected (string) "test-toString" but got (undefined) undefined
FAIL link.nonce: setAttribute() to object "test-valueOf" assert_equals: IDL get expected (string) "test-valueOf" but got (undefined) undefined
FAIL link.nonce: IDL set to "" assert_equals: getAttribute() expected "" but got "test-valueOf"
FAIL link.nonce: IDL set to " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo " assert_equals: getAttribute() expected " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo " but got "test-valueOf"
FAIL link.nonce: IDL set to undefined assert_equals: getAttribute() expected "undefined" but got "test-valueOf"
FAIL link.nonce: IDL set to 7 assert_equals: getAttribute() expected "7" but got "test-valueOf"
FAIL link.nonce: IDL set to 1.5 assert_equals: getAttribute() expected "1.5" but got "test-valueOf"
FAIL link.nonce: IDL set to true assert_equals: getAttribute() expected "true" but got "test-valueOf"
FAIL link.nonce: IDL set to false assert_equals: getAttribute() expected "false" but got "test-valueOf"
FAIL link.nonce: IDL set to object "[object Object]" assert_equals: getAttribute() expected "[object Object]" but got "test-valueOf"
FAIL link.nonce: IDL set to NaN assert_equals: getAttribute() expected "NaN" but got "test-valueOf"
FAIL link.nonce: IDL set to Infinity assert_equals: getAttribute() expected "Infinity" but got "test-valueOf"
FAIL link.nonce: IDL set to -Infinity assert_equals: getAttribute() expected "-Infinity" but got "test-valueOf"
FAIL link.nonce: IDL set to "\0" assert_equals: getAttribute() expected "\0" but got "test-valueOf"
FAIL link.nonce: IDL set to null assert_equals: getAttribute() expected "null" but got "test-valueOf"
FAIL link.nonce: IDL set to object "test-toString" assert_equals: getAttribute() expected "test-toString" but got "test-valueOf"
FAIL link.nonce: IDL set to object "test-valueOf" assert_equals: IDL get expected (string) "test-valueOf" but got (object) object "test-valueOf"
PASS link.integrity: 32 tests
PASS link.hreflang: 32 tests
PASS link.type: 32 tests
PASS link.referrerPolicy: 27 tests
FAIL link.referrerPolicy: setAttribute() to "same-origin" assert_equals: IDL get expected "same-origin" but got ""
PASS link.referrerPolicy: 3 tests
FAIL link.referrerPolicy: setAttribute() to "SAME-ORIGIN" assert_equals: IDL get expected "same-origin" but got ""
PASS link.referrerPolicy: 5 tests
FAIL link.referrerPolicy: setAttribute() to "strict-origin" assert_equals: IDL get expected "strict-origin" but got ""
PASS link.referrerPolicy: 3 tests
FAIL link.referrerPolicy: setAttribute() to "STRICT-ORIGIN" assert_equals: IDL get expected "strict-origin" but got ""
PASS link.referrerPolicy: 5 tests
FAIL link.referrerPolicy: setAttribute() to "strict-origin-when-cross-origin" assert_equals: IDL get expected "strict-origin-when-cross-origin" but got ""
PASS link.referrerPolicy: 3 tests
FAIL link.referrerPolicy: setAttribute() to "STRICT-ORIGIN-WHEN-CROSS-ORIGIN" assert_equals: IDL get expected "strict-origin-when-cross-origin" but got ""
PASS link.referrerPolicy: 30 tests
FAIL link.referrerPolicy: IDL set to "same-origin" assert_equals: IDL get expected "same-origin" but got ""
PASS link.referrerPolicy: 3 tests
FAIL link.referrerPolicy: IDL set to "SAME-ORIGIN" assert_equals: IDL get expected "same-origin" but got ""
PASS link.referrerPolicy: 5 tests
FAIL link.referrerPolicy: IDL set to "strict-origin" assert_equals: IDL get expected "strict-origin" but got ""
PASS link.referrerPolicy: 3 tests
FAIL link.referrerPolicy: IDL set to "STRICT-ORIGIN" assert_equals: IDL get expected "strict-origin" but got ""
PASS link.referrerPolicy: 5 tests
FAIL link.referrerPolicy: IDL set to "strict-origin-when-cross-origin" assert_equals: IDL get expected "strict-origin-when-cross-origin" but got ""
PASS link.referrerPolicy: 3 tests
FAIL link.referrerPolicy: IDL set to "STRICT-ORIGIN-WHEN-CROSS-ORIGIN" assert_equals: IDL get expected "strict-origin-when-cross-origin" but got ""
PASS link.referrerPolicy: 5 tests
PASS link.charset: 32 tests
PASS link.rev: 32 tests
PASS link.target: 32 tests
PASS meta.title: 32 tests
PASS meta.lang: 32 tests
PASS meta.dir: 62 tests
PASS meta.className (<meta class>): 32 tests
PASS meta.hidden: 33 tests
PASS meta.accessKey: 32 tests
PASS meta.tabIndex: 24 tests
PASS meta.name: 32 tests
PASS meta.httpEquiv (<meta http-equiv>): 32 tests
PASS meta.content: 32 tests
PASS meta.scheme: 32 tests
PASS style.title: 32 tests
PASS style.lang: 32 tests
PASS style.dir: 62 tests
PASS style.className (<style class>): 32 tests
PASS style.hidden: 33 tests
PASS style.accessKey: 32 tests
PASS style.tabIndex: 24 tests
PASS style.media: 32 tests
PASS style.type: 32 tests
Harness: the test ran to completion.
